feat: problem 136, 137, 139

+++ src/problem_0091/mod.rs
@@ -10,12 +10,12 @@ fn num_decodings(s: &str) -> i32 {
     }
 
     for i in 1..s.len() {
-        let current = bytes[i] as u8 - b'0';
+        let current = bytes[i] - b'0';
         if current != 0 {
             dp[i] += dp[i - 1];
         }
 
-        let previous = bytes[i - 1] as u8 - b'0';
+        let previous = bytes[i - 1] - b'0';
 
         if (10..=26).contains(&(previous * 10 + current)) {
             dp[i] += if i >= 2 { dp[i - 2] } else { 1 }
